backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
defusedxml==0.8.0rc2
Django==4.2.16
django-environ==0.11.2
django-filter==24.3
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
gunicorn==23.0.0
idna==3.10
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.9.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
